model/requirements.txt
absl-py==0.9.0
astor==0.8.1
cycler==0.10.0
gast==0.3.2
grpcio==1.26.0
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.0
numpy==1.18.1
opencv-python==4.1.0.25
Pillow==7.0.0
protobuf==3.11.2
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0   
termcolor==1.1.0
Werkzeug==0.16.0
